update production and machine info

+++ CIM.Model/ProductionInfoModel.cs
@@ -35,7 +35,7 @@ namespace CIM.Model
         public string Description { get; set; }
         public string ShopNo { get; set; }
         public int Target { get; set; }
-        public int Defect { get; set; }
+        public decimal Defect { get; set; }
         public void ResetMachineInfo(int machineId)
         {
             MachineId = machineId;
